+++ mayan/apps/document_signatures/models.py
@@ -12,11 +12,15 @@ from mayan.apps.databases.model_mixins import ExtraDataModelMixin
 from mayan.apps.django_gpg.exceptions import VerificationError
 from mayan.apps.django_gpg.models import Key
 from mayan.apps.documents.models import DocumentFile
-from mayan.apps.events.classes import EventManagerSave
+from mayan.apps.events.classes import (
+    EventManagerMethodAfter, EventManagerSave
+)
 from mayan.apps.events.decorators import method_event
 from mayan.apps.storage.classes import DefinedStorageLazy
 
-from .events import event_detached_signature_uploaded
+from .events import (
+    event_detached_signature_deleted, event_detached_signature_uploaded
+)
 from .literals import STORAGE_NAME_DOCUMENT_SIGNATURES_DETACHED_SIGNATURE
 from .managers import DetachedSignatureManager, EmbeddedSignatureManager
 
@@ -167,6 +171,11 @@ class DetachedSignature(ExtraDataModelMixin, SignatureBaseModel):
     def __str__(self):
         return '{}-{}'.format(self.document_file, _('signature'))
 
+    @method_event(
+        event_manager_class=EventManagerMethodAfter,
+        event=event_detached_signature_deleted,
+        target='document_file'
+    )
     def delete(self, *args, **kwargs):
         signature_file_name = self.signature_file.name
         if signature_file_name: