+++ mayan/apps/document_signatures/tests/test_api.py
@@ -11,8 +11,8 @@ from mayan.apps.documents.tests.mixins.document_mixins import DocumentTestMixin
 from mayan.apps.rest_api.tests.base import BaseAPITestCase
 
 from ..events import (
-    event_detached_signature_created, event_detached_signature_uploaded,
-    event_embedded_signature_created
+    event_detached_signature_created, event_detached_signature_deleted,
+    event_detached_signature_uploaded, event_embedded_signature_created
 )
 from ..permissions import (
     permission_document_file_sign_detached,
@@ -25,13 +25,13 @@ from ..permissions import (
 from .literals import TEST_SIGNED_DOCUMENT_PATH
 from .mixins import (
     DetachedSignatureAPIViewTestMixin, EmbeddedSignatureAPIViewTestMixin,
-    SignatureTestMixin
+    DetachedSignatureTestMixin
 )
 
 
 class DetachedSignatureDocumentAPIViewTestCase(
     DocumentTestMixin, DetachedSignatureAPIViewTestMixin,
-    KeyTestMixin, SignatureTestMixin, BaseAPITestCase
+    KeyTestMixin, DetachedSignatureTestMixin, BaseAPITestCase
 ):
     auto_upload_test_document = False
 
@@ -76,7 +76,12 @@ class DetachedSignatureDocumentAPIViewTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self.test_document_file)
+        self.assertEqual(events[0].target, self.test_document_file)
+        self.assertEqual(events[0].verb, event_detached_signature_deleted.id)
 
     def test_trashed_document_signature_detached_delete_api_view_with_access(self):
         self._upload_test_document()